+++ client/src/fetchtrack.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+import FetchRecommendations from 'server/controllers'
 
 //backend function variables
 const apiURL = "https://api.reccobeats.com/v1"
@@ -13,12 +14,14 @@ export default function ShowPlaylist(){
     const [tracks, setTracks] = useState(null);
 
     // TODO: rewrite this to send request to api endpoint, which is in server side
-    useEffect(() => {
-        FetchRecommendations().then(fetchedTracks =>{
-            setTracks(fetchedTracks.content);
-            setLoading(false);
-        })
-    }, []);
+    // useEffect(() => {
+    //     FetchRecommendations().then(fetchedTracks =>{
+    //         setTracks(fetchedTracks.content);
+    //         setLoading(false);
+    //     })
+    // }, []);
+
+    
 
     if (loading){
         // placeholder loader
@@ -86,54 +89,3 @@ function GenerateEmbedURL(track){
 
 // TODO: Add the below functions to backend later on
 // FUTURE BACKEND FUNCTIONS
-async function FetchRecommendations(seeds = ['d58affe1-3e80-4318-b33f-9f85bbecf693'], size = 5, features = {energy:"", valence:""}){
-    //test url: https://api.reccobeats.com/v1/track/recommendation?size=3&seeds=83dc71c7-b9da-466b-a198-bb3c29ee8f00
-    const seedsString = SeedsToString(seeds);
-
-    let recsURL = `/track/recommendation?size=${size}&seeds=${seedsString}`;
-    let featuresString = FeaturesToString(features);
-    
-    try{
-        const response = await fetch(apiURL+recsURL+featuresString);
-
-        if (!response.ok){
-            throw new Error("Error: could not fetch recommendations");
-        }
-
-        return await response.json();
-    }
-
-    catch (error){console.log(error)}
-}
-
-function SeedsToString(seeds){
-    let output = "";
-
-    for (let x = 0; x < seeds.length; x++){
-        if (x > 0){
-            output += ",";
-        }
-
-        output += seeds[x];
-    }   
-
-    return output;
-}
-
-function FeaturesToString(obj){;
-    let featuresString = "";
-
-    for (x in audioFeatures){
-        const feature = audioFeatures[x]
-        const value = obj[feature];
-
-        // if value exists
-        if (value){
-            string+= `&${feature}=${value}`;
-        }
-    }
-
-    return featuresString;
-}
-
-
